requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
aiohttp>=3.9.5
tqdm>=4.66.4
requests>=2.32.3
python-dotenv>=1.0.1
flask>=3.0.3
